requirements.txt
boto3
praw
colorific
pillow
requests
sqlalchemy